add goal based planning
This commit adds goal based planning. You can define a goal as a desired
piece of state and, with the appropriate goal to methods defined, the
system will attempt to build a plan to mutate the state to match that goal.

```perl
    my $state = {...};
    my $goal = { loc => { alice => 'park', bob => 'park' } };

    my $planner = PLHop::Planner->new(
        domain    => $domain,
        state     => $state,
        todo_list => [ $goal ],
    );

    my @plan = $planner->plan();
```

NOTE: Internally this very much currently expects on the goals being a
hashref. Future work will need to be done to turn goals into
objects/interfaces that can be iterated on. Likely we'll want to
transform State into a similar object so that there is syncronicity.